+++ kogitoq/gadaq/src/main/java/life/genny/gadaq/search/FilterGroupService.java
@@ -1072,12 +1072,14 @@ public class FilterGroupService {
             boolean coded = isCode(value);
             String dropdownVal =  getDropdownValue(value);
             String sbe =  filterService.getSbeTableFromCache();
+            List<String> definitions = filterService.getListDefinitionCodes(sbe);
+
             if(isBucketSbe(sbe)){
                 // search bucket
-                filterService.handleQuickSearchDropdownByBucket(dropdownVal, coded);
+                filterService.handleQuickSearchDropdownByBucket(dropdownVal,coded);
             }else {
                 // search table
-                filterService.handleQuickSearchDropdown(dropdownVal, coded, sbe);
+                filterService.handleQuickSearchDropdown(dropdownVal, coded, sbe,definitions);
             }
 
         } catch (Exception ex){
@@ -1112,7 +1114,7 @@ public class FilterGroupService {
             String value = msg.getData().getValue();
             if(!value.isEmpty()) {
                 String sbeCode = filterService.getSbeTableFromCache();
-                List<String> definitions = getListDefinitionCodes(sbeCode);
+                List<String> definitions = filterService.getListDefinitionCodes(sbeCode);
 
                 filterService.sendListQuickSearches(Question.QUE_QUICK_SEARCH_GRP,Question.QUE_QUICK_SEARCH,
                         SearchEntity.SBE_QUICK_SEARCH, Attribute.PRI_NAME,FilterConst.VALUE,value,definitions);
@@ -1131,60 +1133,5 @@ public class FilterGroupService {
         return SearchEntity.SBE_PROCESS.equals(code);
     }
 
-    /**
-     * Return the list of base entity code
-     * @param sbeCode Search base entity code
-     * @return List of base entity code
-     */
-    public List<String> getListDefinitionCodes(String sbeCode) {
-        List<String> definitions = new ArrayList<>();
 
-        // bucket page
-        if(SearchEntity.SBE_PROCESS.equals(sbeCode)) {
-            addDefinitionCodeByBucket(definitions);
-        // Table
-        } else {
-            String defCode = getDefinitionCode(sbeCode);
-            if(!defCode.isEmpty()) {
-                definitions.add(defCode);
-            }
-        }
-
-        return definitions;
-    }
-
-    /**
-     * Return definition code
-     * @param sbeCode search base entity code
-     * @return definition code
-     */
-    public String getDefinitionCode(String sbeCode) {
-        SearchEntity search = CacheUtils.getObject(userToken.getProductCode(),sbeCode,SearchEntity.class);
-        List<ClauseContainer> clauses = search.getClauseContainers();
-        for(ClauseContainer clause : clauses) {
-            if(clause.getFilter().getCode().equals(Attribute.LNK_DEF)) {
-                return clause.getFilter().getValue().toString();
-            }
-        }
-        return "";
-    }
-
-    /**
-     * Add definition to the list of definition codes
-     * @param definitions List of definition codes
-     */
-    public void addDefinitionCodeByBucket(List<String> definitions) {
-        PCM pcm = beUtils.getPCM(PCM.PCM_PROCESS);
-
-        List<EntityAttribute> locations = pcm.findPrefixEntityAttributes(Prefix.LOCATION);
-        for (EntityAttribute entityAttribute : locations) {
-            String value = entityAttribute.getAsString();
-            if (value.startsWith(Prefix.SBE)) {
-                String defCode = getDefinitionCode(value);
-                if(!defCode.isEmpty()) {
-                    definitions.add(defCode);
-                }
-            }
-        }
-    }
 }
